+++ lib/yt/models/file_detail.rb
@@ -12,14 +12,8 @@ module Yt
         @data = options[:data] || {}
       end
 
-      # @return [Integer] the file size of the video (in bytes).
       has_attribute :file_size, type: Integer
-
-      # @return [String] the file type of the file. May be one of:
-      #   archive, audio, document, image, other, project, or video.
       has_attribute :file_type
-
-      # @return [String] the container of the video (e.g. 'mov').
       has_attribute :container
     end
   end